requirements.txt
fastapi
uvicorn
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
streamlit
